stateHelpers: move ToggleBypassFX to state_helpers
