+++ BryanChi_FX_Devices/src/helpers/state_helpers.lua
@@ -1,6 +1,6 @@
-local state_heplers = {}
+local state_helpers = {}
 
-function state_heplers.DeleteAllParamOfFX(FXGUID, TrkID)
+function state_helpers.DeleteAllParamOfFX(FXGUID, TrkID)
     for p, _ in pairs(FxdCtx.Trk.Prm.FXGUID) do
         if FxdCtx.Trk.Prm.FXGUID[p] == FXGUID and FXGUID ~= nil then
             FxdCtx.Trk.Prm.Inst[TrkID] = FxdCtx.Trk.Prm.Inst[TrkID] - 1
@@ -14,4 +14,17 @@ function state_heplers.DeleteAllParamOfFX(FXGUID, TrkID)
     end
 end
 
-return state_heplers
+---@param LT_Track MediaTrack
+---@param FX_Idx integer
+function state_helpers.ToggleBypassFX(LT_Track, FX_Idx)
+    FxdCtx.FX.Enable = FxdCtx.FX.Enable or {}
+    FxdCtx.FX.Enable[FX_Idx] = r.TrackFX_GetEnabled(LT_Track, FX_Idx)
+    if FxdCtx.FX.Enable[FX_Idx] == true then
+        r.TrackFX_SetEnabled(LT_Track, FX_Idx, false)
+    elseif FxdCtx.FX.Enable[FX_Idx] == false then
+        r.TrackFX_SetEnabled(LT_Track, FX_Idx, true)
+    end
+end
+
+
+return state_helpers
